bot026_keyword500_gpt_test/requirements.txt
pandas==2.0.3
openpyxl==3.1.2
openai==0.28.0
python-dotenv==1.0.0 